feat(dashboard): add 3D campus visualization components and zone management
- Implemented DashboardHelpers for utility functions like clamp and deriveStatus.
- Created DashboardScene for rendering the 3D campus layout with zones and buildings.
- Developed DashboardSidebar for zone selection and status display.
- Defined DashboardTypes for zone data structure and status management.
- Added FirstPersonController for user navigation in the 3D space.
- Built GroundComponent for the campus ground and playground representation.
- Developed RoadsComponent for rendering roads and markings in the campus layout.
